fix(dialog): don't move focus to dialog container if focus is already inside the dialog
This is a follow-up to https://github.com/angular/components/pull/16221. What we hadn't accounted for in the aforementioned PR is that the consumer may have turned off `autoFocus` so that they can handle it themselves and with our changes focus would be reset back to the container. These changes add an extra check which will ensure that focus is only moved if it's not inside the dialog already.