build: fix snapshot local browsers cronjob test
The `snapshot_tests_local_browsers` job currently seems
to fail because in order to run the setup snapshots script,
we already need to have the `node_modules` installed. This
means that we need to run Yarn twice for the cronjob. Though
this shouldn't have any performance implications because
we recently updated Yarn to a version that is able to
properly cache/restore the `node_modules`.