fix(ng-update): do not rely on node-glob for finding rule directories (#16381)
We currently use `glob` in order to find rule directories automatically. This
has implications because it means that we depend on `glob` that is usually
brought in transitively by other project dependencies. This means that the
glob version is not guaranteed to be working due to previous `glob`
issues on windows.

Related to #16208