test(material-experimental/chips) Port unit tests for mdc chips
Ported existing unit tests as follows:
- Copied chip-remove.spec.
- Copied chip-input.spec and changed its test component template to use
grid/row.
- Copied chip-list.spec basic tests to chip-set.spec.
- Copied chip-list.spec to chip-listbox.spec and removed form-field
integration, mat-chip-remove integration, and related keyboard shortcuts.
- Copied chip-list.spec to chip-grid.spec and removed the selection behavior.
Updated keyboard tests to reflect that it is using GridFocusKeyManager,
which has both an active row and an active column. Updated some of the form
tests where the value was being changed via selection to change the value
in a different way, like by adding a chip.
- Copied chip.spec basic tests to chip.spec.
- Copied chip.spec to chip-option.spec and removed logic related to user
removal of the chip.
- Copied chip.spec to chip-row.spec and removed selection logic.

Also updated component code to fix bugs/missing features that were caught by
the unit tests:
- Added missing ControlValueAccessor method setDisabledState to mat-chip-listbox
and mat-chip-grid
- Stopped overwriting  user-defined tab index to mat-chip-listbox and mat-chip-grid.
- Removed role from empty mat-chip-listbox and mat-chip-grid.
- Added a new subscription in mat-chip-set to chip destroyed events, and moved the
logic that updates the lastDestroyedChipIndex out of the removed subscription into
the new destroyed subscription.
- Replaced _optionChip and _rowChips ViewChildren in mat-chip-listbox and
mat-chip-grid with just overriding _chips, to fix bugs where the
_chips.changes subscription was trying to do things with
_optionChips/_rowChips but those QueryLists were still out of date.
- Added placeholder setter to mat-chip-grid.
- Added value setter to mat-chip-listbox.
- Added injected animationMode to mat-chip and its subcomponents.
- Updated rippleConfig to be injected into mat-chip and its subcomponents.
- Switched click handler to mousedown handler in mat-chip-row.
- Added mat-chip-remove class to remove icon.