fix(expansion): incorrect margin for "before" indicator when expanded
Fixes the panel header indicator not being positioned correctly when it's set to the `before` position and it's expanded. The issue comes from the fact that we were using `padding` instead of `margin` which changes the center of the rotation.

Also fixes that the `before` position of the indicator doesn't account for RTL.