test: rework flaky tests that depend on getComputedStyle
Currently most of the grid list tests as well as one of tab nav bar test are flaky, because we use `getComputedStyle` for assertions and the return values can be formatted differently depending on the browser. These changes move them to either pick up values directly from the DOM or to measure the element with `getBoundingClientRect`.