refactor(button): render focus tint and ripples behind button content
Reworks the button focus indication and ripples so that they render behind the text of the button, instead of top of it. This helps with the contrast ratios while focused.

Also gets rid of the `mat-button-focus-overlay` element in favor of using `::before` which saves us the extra DOM node.

Incorporates some of the changes from #11961 so that we don't have to increase the size of the CSS selectors.