build: allow inspecting test nodejs process w/ bazel
Adds a new `debug` configuration for Bazel that can be used to
open the Chrome Devtools Debugger in order to debug a bazel
test that runs with NodeJS. This is similar to what `angular/angular`
provides.

*Note*: This does not make debugging for web test suites easier. For
web test suites we need to find a better solution. I have something in mind
where developers can run `bazel run` instead of `bazel test` to open the
karma test suite in non-capture mode (allowing easy debugging)