fix(ng-update): material update fails due to circular dependency
Currently running `ng update` for both Material and CDK fails because there
is a circular dependency in the CDK schematics that cause the update-tool
exports to be not available to the Material ng-update schematic.

Without putting to much effort into ensuring no circular dependencies, this
commit also sets up Madge to avoid such issues in the future.