fix(slide-toggle): invalid required validator in template-driven forms
Currently using the `slide-toggle` in a form using template-driven forms causes
the slide-toggle to retrieve a wrong validator if the `required` attribute is set.

This is because by default `@angular/forms` uses an input validator
that ensures that the value is just defined. This is always the case for
a slide-toggle since the value is always `true` or `false`.

The solution to this problem is that we need to provide the checkbox validator
for required slide-toggle components. The checkbox validator from the forms
package ensures that the control is only valid if the slide-toggle is checked.