style(mdc chips) Fix clashing input and remove classes
We have slightly different styles for the MDC versions of
mat-chip-remove and mat-chip-input. Projects that import
all angular material css are also getting the non-mdc chips
css, which is applying the non-mdc styles to these elements.

Add -mdc to mat-chip-remove and mat-chip-input classes in the
mdc module to differentiate them. All the other mdc chip
classes already contain -mdc for the same reason.