docs(datepicker): modules to import to start using
The datepicker requires two modules to be able to be used. It requires a
datepicker *adapter* as well.

This is easy to miss when scrolling
through the docs and seeing the example code. Adapters should
be mentioned at least once in
the intro to make the reader
aware that they are a requirement.