fix(menu): keyboard controls not working if all items are disabled
Currently we depend on `focusFirstItem` to bring focus into the menu, however if all the items are disabled, focus will stay on the trigger and the user won't get feedback that something has happened. Furthermore, the keyboard shortcuts that are implemented in the menu won't work either, because the keyboard event listener is on the menu.

These changes work around the issue by setting focus on the `role="menu"` if none of the items were focusable.

Fixes #16565.