test(sidenav): refactor flaky tests in Edge (#16625)
Refactors some tests to (hopefully) be less flaky in Edge. In two of them I've replaced calls to `getComputedStyle` to a check against a class in the DOM. On other one I've changed the assertions for IE and Edge to something that is less accurate, but also less likely to flake.

(cherry picked from commit 8e321aee72740ee8271c193d3872866a6e980a40)