fix(slider): inconsistent cursor behavior if user's pointer moves around while dragging
Currently we only change the cursor in the slider if the user's pointer is over the thumb or thumb label. This doesn't work very well while dragging, because the elements aren't very large and the pointer could move off of it temporarily while dragging. These changes make it so the cursor for the entire slider is changed while dragging.

Furthermore, on Safari the cursor goes into text selection mode while dragging. These changes address this issue as well.

Fixes #14613.