fix(slider): displayWith function never called with "null"
Currently the "displayWith" function for a `MatSlider` is typed
in a way that implies that the function is sometimes called with
null as value. This is never the case because the `value` property
of the slider always returns a number but is just typed as `null`-able
because getters and setters cannot have different types.

This fixes the type for the `displayWith` function as it will never be
called with `null`.