feat(material-experimental): add option to filter checkbox harnesses by name
Adds an option to the checkbox harnesses that allows developers
to filter harnesses by checkbox names. This is helpful because
developers can not filter checkboxes by name using CSS selectors
because the `name` input binding is not bound to the host DOM element.

Similar to https://github.com/angular/components/pull/16609#discussion_r308323723