feat(material-experimental): add menu test harness skeleton (#16615)
Adds a skeleton for the menu test harness. It currently captures the
basic functionality for the trigger. The functionality for the menu
panel/items will come in a follow-up PR. As part of this, I'm adding a
css class to the menu trigger so that it's findable via DOM.

(cherry picked from commit 3857ab262cc816f2ff3822cb629524ad4de079d5)