minor tweak to mdc chips adapter for latest upstream update to unbreak pantheon
angular already manages focus fully and this hook does negligible good and a lot of harm (e.g., focusing thigns that were never meant to be focused). long term we probably want to consolidate the code, but for now just keep it unbroken.

squash this commit into the main PR