fix(tabs): tabs with background overriding ripple color for all descendants
When a tab group has a background color we need to change its ripple color so it doesn't blend in with the background, however the current selector is too broad which means that it also targets all the ripples inside the tabs content. These changes narrow down the selector to avoid the issue.