ci: temporarily disable gesture config tests
We need to temporarily disable the gesture tests in order to
land a PR in framework that makes Hammer support optional
(https://github.com/angular/angular/pull/32203). Otherwise,
the Material/FW integration tests will fail because Material
modules haven't yet opted into Hammer.

Unfortunately, we cannot yet fix this properly by adding the
HammerModule to Material because Material is not using a
release that has HammerModule. Once the PR lands and Material
updates to a 9.0.0-next release, we can add the HammerModule
properly and turn these tests back on.