build: add blocklist to disable individual tests in the project
Similarly to what we had in the `ivy-2019` branch for disabling
individual tests, we now need something similar that works with
Bazel. This is necessary because there are still breaking changes
that could land in the framework repository and break individual
tests of the component repository.

Framework now needs a way to disable those specific tests until
the components repository migrated the broken tests (this can
be delayed by weeks since the breaking change is not necessarily
yet released)