fix(cdk/a11y): FocusMonitor emits rogue values in IE 11
Works around a bug in Internet Explorer where the `focus` and `blur`
events do not have the `relatedTarget` set which we rely on to determine
whether focus is leaving the container we are monitoring. Uses `focusin`
and `focusout` events instead as these behave the same but get the
`relatedTarget` set in all our supported browsers.

Fixes #15517