build: re-enable disabled gesture config test
eaf70ca2a0600757041633976b29ab5a95d08296 temporarily disabled the gesture
config tests because the blocklist didn't work. Now that the blocklist works on
the framework side, we can re-enable this test again.

Once this PR lands, we can update the SHA on the framework repo and
add the gesture tests to the blocklist.