fix(drag-drop): connected drop zones not working inside shadow root
Fixes not being able to drop into a connected drop list when using `ShadowDom` view encapsulation. The issue comes from the fact that we use `elementFromPoint` to figure out whether the user's pointer is over a drop list. When the element is inside a shadow root, calling `elementFromPoint` on the `document` will return the shadow root.

These changes fix the issue by calling `elementFromPoint` from the shadow root instead.

Fixes #16898.