build: add dev app proxy server for testing across devices
We currently can't test the dev app across devices, because the Bazel server doesn't allow us to configure the server host.

These changes adds a very simple, temporary, proxy server that runs in parallel to Bazel dev server and proxies all requests to it, while allowing devices to connect to it.