fix(material-experimental): mdc-typography should use default typography config
Currently if someone replaces the `angular-material-typography` mixin with the
`angular-material-typography-mdc` mixin, the behavior will be different if no explicit
typography config is passed. This is because the `mdc-typography` mixin does not
use the default `mat-typography-config` if no config is specified.

To make it consistent with the behavior of the `angular-material-typography` mixin, the
fallback for the default typography config will be added to the mixin.