feat(sort): remove hammer dependency
These changes remove the dependency on Hammer.js from the `matSortHeader`. I decided to remove the `longpress` gesture, rather than implement it with vanilla JS, because it doesn't look like it's worth the trouble. The gesture was meant to show the hover indication if the user holds their pointer on the header, however the touch area is so small that it's most likely that the user won't be able to see the hover indication anyway, because it'll be under their finger. Furthermore, there's nothing indicating to the user that they can long press this element anyway.