build: do not include material prebuilt theme twice in legacy tests
Apparently we include the Angular Material prebuilt theme twice in the
legacy Karma tests. This is because we also generate a second folder
called `esm5` in the package output (for later use in the release
composing step). This folder also contains the prebuilt themes as
a side-effect of copying the assets for esm5 inlining/bundling.

In any case, we should not import the theme using a glob, but rather
use an explicit path (similar to how we do it for the mdc theme).