fix(progress-spinner): animation node inserted into wrong style root when using ngIf with ShadowDom encapsulation
If the progress spinner is inside the shadow DOM, we attach the `style` tag with the animation to the shadow root, however since we do the insertion inside the constructor, the spinner might not have been moved into the shadow root yet. These changes move the logic into `ngOnInit` which fires late enough for the node to be in the correct place.