fix: annotate a couple base classes with @Directive for Ivy
There are two base classes in Material/CDK which are intended to be extended
by user code: BasePortalHost and MatFormFieldControl.

This commit annotates them with `@Directive`. The best way to do this is to
use `@Directive()` but this is not supported by Angular 8.x, on which
Material and the CDK still depend. Therefore, a workaround is used with a
selector which will never be matched. Modules are also added to satisfy the
View Engine compiler.