build: setup primary entry-point package.json files to match release output
The primary entry-point `package.json` files are not generated by the
`ng_package` bazel rule. This is by intention because it does not
try to overwrite existing `package.json` files in the release output
(i.e. to preserve placeholders; dependencies etc.)

We need to update the `package.json` fields to match the new
release output that is generated by Bazel. Also adds a new
release output validation that ensures that those fields in the
`package.json` files are resolving to existing files.