build: run custom tslint rules on new packages (#17033)
Apparently we didn't run the custom tslint rules for the new packages
that have been merged (`youtube-player` and `google-maps`). This is
because we have a whitelist of packages we want to run these custom
lint rules against.

It should be actually the other way around where we only specify the folders
we _don't_ want to lint against. That way, new folders are automatically linted
and we need to explicitly exclude them if we don't want to run for those.

(cherry picked from commit 0d7e0383a61295eb624beed02bc81181c8af6336)