build: fix npm package for google-maps
Apparently the `google-maps` package has the same problem
as the `material-examples` package where the metadata for
the Bazel sub-packages is not incldued in the entry-point of
the `@angular/google-maps` entry-point. There are two solutions
to make it work with Angular package format:

* Making the packages generate flat module bundles and including
them as secondary entry-points
* Only having one Bazel package so that metadata will be included
in the flat module bundle.

I chose the latter as the `google-maps` package is still very small
and doesn't have a lot of dependencies, so in terms of speed it is
even faster to have on package. Also it seems like there were never
plans to have secondary entry-points for that package.

*Note*: we need this commit as part of the docs-content fix because
the google-maps package runs on CI and will fail as it misses various
`public-api.ts` files for the flat module bundles.