build: do not build examples package through gulp
We no longer can build the examples package through gulp. This is
because we did not add individual `tsconfig-build.json` files to
the nested entry-points when we did the restructuring of the examples.

This now breaks as the primary entry-point uses module name
imports as of the docs-content metadata fix. It's okay to not build the
examples package through gulp in the `8.2.x` branch as we don't
need it in the gulp setup anyway.