build: fix docs-content release output (#17125)
* build: add workaround for invalid imports generation in bazel

Applies a workaround for the invalid imports generation issue
in the Bazel `ng_module` rule. Read more about this here:

https://hackmd.io/@devversion/ryCOwIKIS.

* build: fix missing metadata for example modules

Fixes the missing metadata for example modules. This happens because
we built the individual example modules without a flat module bundle. Not
building with a flat module bundle means that metadata will not be generated
in the `ng_module` bazel rule.

* build: workaround for creating ng_package on windows

Workaround for https://github.com/angular/angular/issues/32603

* build: fix invalid rollup globals for material experimental package

* build: do not generate flat module bundles for testonly targets

This is necessary as `ng_module` rules with a `module_name` either
need a `public-api.ts` or `module.ts` file.

* build: fix npm package for google-maps

Apparently the `google-maps` package has the same problem
as the `material-examples` package where the metadata for
the Bazel sub-packages is not incldued in the entry-point of
the `@angular/google-maps` entry-point. There are two solutions
to make it work with Angular package format:

* Making the packages generate flat module bundles and including
them as secondary entry-points
* Only having one Bazel package so that metadata will be included
in the flat module bundle.

I chose the latter as the `google-maps` package is still very small
and doesn't have a lot of dependencies, so in terms of speed it is
even faster to have on package. Also it seems like there were never
plans to have secondary entry-points for that package.

*Note*: we need this commit as part of the docs-content fix because
the google-maps package runs on CI and will fail as it misses various
`public-api.ts` files for the flat module bundles.

* chore: fix various examples missing button module

Looks like a few examples were missing the `MatButtonModule` after
the restructuring of the examples. I've noticed these while manually
testing the Bazel release output in the docs app.

* build: do not build examples package through gulp

We no longer can build the examples package through gulp. This is
because we did not add individual `tsconfig-build.json` files to
the nested entry-points when we did the restructuring of the examples.

This now breaks as the primary entry-point uses module name
imports as of the docs-content metadata fix. It's okay to not build the
examples package through gulp in the `8.2.x` branch as we don't
need it in the gulp setup anyway.