build: fix bazel devserver not launching
The custom devserver implementation does currently not
work due to a breaking change in the `rules_nodejs` repo.

Apparently a `nodejs_binary` no longer generates an executable
output without either `.bat` or `.sh`. This is something we
relied on in the devserver launcher script. The fix is as simple
as resolving the executable with `.sh` extension as we run the
devserver launcher from within a `sh_binary` already.