build: ensure all entry-points are included in release package
(1) Overhauls the whole release packaging structure in Bazel. Since it is
very verbose to put everything into a single `.bzl` file in the project root,
information for each release package will be stored in the actual package
folder in a file called `config.bzl`.

(2) We no longer use the term `PACKAGE` for entry-points of
packages. It's not correct and easy to confuse with the actual package.

(3) We group entry-points in nested arrays. This has no effect for now, but
it gives us a nicer way of configurating (nested) entry-points. It might become
useful in the future as well as we have information about nested entry-points of
an entry-point.

(4) Revisits the list of all entry-points so that we now include all `*/testing`
entry-points in the release output. Also a few MDC prototypes were not included
in the Bazel material-experimental release output. Same as clipboard in cdk-experimental.

(5) Simplifies the logic to generate the rollup globals (through functions). Also we move the
rollup globals into a separate `.bzl` file. Also rollup globals are now automatically passed
into the `ng_package` rule.. so we don't need to re-declare them for each package.