build: update to chalk version that works in circleci
Updates to a chalk version that works in CircleCI (due to a more recent
dependency on `supports-color`).

Since the latest chalk version brings official typings that do not declare
individual colors as ESM exports, we need to go through the default export.
This is the official expected behavior with `chalk` according to: https://github.com/chalk/chalk/issues/215.