fix(material-experimental): use harnesses for mat-option
Follow-up from #16620 and #16710. Adds a dedicated harness for `mat-option` and `mat-optgroup`.

Note that some of the code is duplicated. This is because we don't have a shared place where to put the harness so that `mat-autocomplete` and `mat-select` don't have to depend on each other. I've intentionally kept the harnesses to only the methods we need, but once we have `experimental/core`, I'll combine them and implement all of the states that are supported by `mat-option`.