build: generate metadata for cdk keycodes
We need to use "ng_module" for building the `keycodes` entry-point since we
want metadata to be generated for keycode constants. This is necessary
because some of these constants are statically used in component/directive
metadata (e.g. chips) and NGC needs to be able to evaluate these at compile time.