fix(cdk-experimental/clipboard): avoid changing layout with temporary textarea
Currently the clipboard service creates a temporary invisible `textarea` from which to copy the text into the clipboard, however since it only has `opacity: 0`, it has the potential of changing the layout in a noticeable way if it is present for longer.

These changes add some extra styles to avoid changing the layout. I also did some small cleanup to make it a bit more robust.