build: populate release version placeholder properly
Currently we always append the SHA of `HEAD` to the
version placeholder. We only want to append the SHA
of `HEAD` to the version if we are publishing snapshot
build artifacts to the Github build repositories.

This also fixes that the SHA is appended twice to
snapshot builds release output. e.g.

https://github.com/angular/cdk-builds/blob/master/package.json#L3.