build: tests should not fail if stylesheets contain quotes
Currently our workaround that allows us to use stylesheets
in the karma web test suites is not very robust and fails if
stylesheets use single quotes. e.g. `@material/linear-progress`:

https://unpkg.com/browse/@material/linear-progress@3.2.0/_mixins.scss

In order to fix this, we use double-quotes and always escape double-quotes
in the CSS content. This is still a bit hacky, but so is the overall workaround to
make it work.