build: do generate invalid semver versions for snapshot builds
Apparently there is a long-standing issue in the snapshot builds
where we generate invalid versions if the snapshot SHA starts with
zero's. In order to fix this, we need to ensure that the commit SHA is
part of a version segment that starts with text (i.e. `-sha-{val}`).

Fixes #17264