fix(stepper): unable to skip step if completed value is overwritten (#15403)
Currently we allow for the `completed` value of a step to be overwritten, however setting it to `true` still won't allow the user to skip the step, because the `interacted` flag won't be flipped if they haven't been on a particular step before.

Fixes #15310.

(cherry picked from commit bebb9ff784d49c7c949363557000fe75c5aa520d)