build: disable sourcemap generation for all sass binaries
We need to disable sourcemap generation for all sass binaries
as otherwise the inline source map comment will be inlined into
component `styles` in the release output.

Technically this is not a big deal for AOT compilation where those
source-map strings are omitted anyway, but for JIT it will contribute in
payload size, and with the current state of Angular Package Format, these
source-maps would point to non-existent files anyway.

We should just not generate these sourcemaps/comments similarly
to our Gulp release output.