fix(drag-drop): disabled state not synced on data binding changes
Usually we only sync up the drop list ref with the directive right before dragging starts, however if the disabled state is set, it can propagate to the child directives and prevent the user from ever starting a search again. These changes make it so whenever the disable state is set it'll get sync immediately.

Also changes a misleading method name.

Fixes #17325.